Updated README
Updated the README.md for first-time users. More will need to be added,
or a different tutorial file for instructions on updating the project.
This will be handled in the next story.

Added package-lock.json to .gitignore. Will need to decide if it should
be packaged with the template project or not.